requirements.txt

openai>=1.37.0
chromadb>=0.5.3
python-dotenv>=1.0.1
pypdf2>=3.0.1
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.3.0
langdetect>=1.0.9
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
rich>=13.7.0
tqdm>=4.66.0
pycryptodome>=3.20.0
